replace dependencies in App
