--- app/src/main/java/ru/dvn/gitapp/data/remote/GitRetrofit.kt
@@ -1,28 +1,0 @@
-package ru.dvn.gitapp.data.remote
-
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
-import retrofit2.converter.gson.GsonConverterFactory
-import ru.dvn.gitapp.data.remote.user.GithubApi
-
-object GitRetrofit {
-    private const val BASE_URL = "https://api.github.com/"
-
-    private val retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
-        .client(buildHttpClient())
-        .build()
-
-    fun getGithubApi(): GithubApi = retrofit.create(GithubApi::class.java)
-
-    private fun buildHttpClient(): OkHttpClient {
-        return OkHttpClient.Builder()
-            .addInterceptor(HttpLoggingInterceptor().apply {
-                level = HttpLoggingInterceptor.Level.BODY
-            }).build()
-    }
-}
