+++ app/src/main/java/ru/dvn/gitapp/ui/users/details/UserDetailsActivity.kt
@@ -64,7 +64,7 @@ class UserDetailsActivity : AppCompatActivity() {
 
     private fun restoreViewModel(nickName: String): UserDetailsContract.ViewModel {
         return lastCustomNonConfigurationInstance as? UserDetailsContract.ViewModel
-            ?: UserDetailsViewModel(app().mainRepository, nickName)
+            ?: UserDetailsViewModel(app().repository, nickName)
     }
 
     private fun renderDetails(userDetails: UserDetails) {
